Move project init utils to api.commands
