+++ tests/streamlit/test_commands.py
@@ -501,7 +501,7 @@ def test_drop_streamlit(mock_connector, runner, mock_ctx):
     assert ctx.get_query() == f"drop streamlit {STREAMLIT_NAME}"
 
 
-@mock.patch("snowcli.api.project_initialisation._create_project_template")
+@mock.patch("snowcli.api.commands.project_initialisation._create_project_template")
 def test_init_streamlit(mock_create_project_template, runner, temp_dir):
     runner.invoke(["streamlit", "init", "my_project3"])
     mock_create_project_template.assert_called_once_with(
